requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
llama-cpp-python==0.3.20
Pillow==10.4.0
